move: 필요없는 클래스 삭제 및 이동
